# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trip Planner API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/airports": {
      "get": {
        "tags": [
          "lookup"
        ],
        "summary": "Airports",
        "operationId": "airports",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 60,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirportOut"
                  },
                  "title": "Response Airports"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/agent/v1/runs/{run_id}/context": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Run Context",
        "description": "The task: trip, routes with ids and date rules, cheapest known prices, and the rules.",
        "operationId": "run_context",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/v1/airports": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Lookup Airports",
        "operationId": "lookup_airports",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 60,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirportOut"
                  },
                  "title": "Response Lookup Airports"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/v1/runs/{run_id}/flight-quotes": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Submit Flight Quotes",
        "description": "Up to 50 prices. Each is accepted, rejected with reasons, or reported as a duplicate.",
        "operationId": "submit_flight_quotes",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteBatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteBatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/v1/runs/{run_id}/activity-suggestions": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Submit Activity Suggestions",
        "description": "Up to 20 things to do. Each is accepted, rejected with reasons, or reported as a duplicate.",
        "operationId": "submit_activity_suggestions",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestionBatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionBatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/v1/runs/{run_id}/notes": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Add Note",
        "operationId": "add_note",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/agent/v1/runs/{run_id}/finish": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Finish Run",
        "operationId": "finish_run",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptedItem": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          }
        },
        "type": "object",
        "required": [
          "index",
          "id",
          "flags"
        ],
        "title": "AcceptedItem"
      },
      "AcceptedSuggestion": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "index",
          "id"
        ],
        "title": "AcceptedSuggestion"
      },
...
      "AgentAck": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "title": "AgentAck"
      },
      "AirportOut": {
        "properties": {
          "iata": {
            "type": "string",
            "title": "Iata"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "country_code": {
            "type": "string",
            "title": "Country Code"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          }
        },
        "type": "object",
        "required": [
          "iata",
          "name",
          "city",
          "country_code",
          "kind"
        ],
        "title": "AirportOut"
      },
...
      "ContextRoute": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "origins": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Origins"
          },
          "destinations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Destinations"
          },
          "trip_type": {
            "type": "string",
            "title": "Trip Type"
          },
          "depart_from": {
            "type": "string",
            "format": "date",
            "title": "Depart From"
          },
          "depart_to": {
            "type": "string",
            "format": "date",
            "title": "Depart To"
          },
          "nights": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nights"
          },
          "return_window": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "date"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Window"
          },
          "passengers": {
            "type": "integer",
            "title": "Passengers"
          },
          "adults": {
            "type": "integer",
            "title": "Adults"
          },
          "children": {
            "type": "integer",
            "title": "Children"
          },
          "cabin": {
            "type": "string",
            "title": "Cabin"
          },
          "max_stops": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Stops"
          },
          "cheapest_known": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cheapest Known"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "origins",
          "destinations",
          "trip_type",
          "depart_from",
          "depart_to",
          "nights",
          "return_window",
          "passengers",
          "adults",
          "children",
          "cabin",
          "max_stops",
          "cheapest_known"
        ],
        "title": "ContextRoute"
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          }
        },
        "type": "object",
        "required": [
          "field",
          "msg"
        ],
        "title": "FieldError"
      },
      "FinishIn": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "partial",
              "failed"
            ],
            "title": "Status"
          },
          "summary": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary"
          },
          "sources_checked": {
            "items": {
              "type": "string",
              "maxLength": 300,
              "minLength": 0
            },
            "type": "array",
            "maxItems": 40,
            "title": "Sources Checked"
          },
          "issues": {
            "items": {
              "type": "string",
              "maxLength": 500,
              "minLength": 0
            },
            "type": "array",
            "maxItems": 20,
            "title": "Issues"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "summary"
        ],
        "title": "FinishIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NoteIn": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Body"
          },
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Urls"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "body"
        ],
        "title": "NoteIn"
      },
      "NoteOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "trip_id": {
            "type": "integer",
            "title": "Trip Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Urls"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "trip_id",
          "run_id",
          "title",
          "body",
          "urls",
          "created_at"
        ],
        "title": "NoteOut"
      },
...
      "QuoteBatchIn": {
        "properties": {
          "quotes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Quotes"
          }
        },
        "type": "object",
        "required": [
          "quotes"
        ],
        "title": "QuoteBatchIn"
      },
      "QuoteBatchResult": {
        "properties": {
          "accepted": {
            "items": {
              "$ref": "#/components/schemas/AcceptedItem"
            },
            "type": "array",
            "title": "Accepted"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedItem"
            },
            "type": "array",
            "title": "Rejected"
          },
          "duplicates": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Duplicates"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "rejected",
          "duplicates"
        ],
        "title": "QuoteBatchResult"
      },
...
      "RejectedItem": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "index",
          "errors"
        ],
        "title": "RejectedItem"
      },
...
      "RunContext": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "today": {
            "type": "string",
            "format": "date",
            "title": "Today"
          },
          "trip": {
            "additionalProperties": true,
            "type": "object",
            "title": "Trip"
          },
          "routes": {
            "items": {
              "$ref": "#/components/schemas/ContextRoute"
            },
            "type": "array",
            "title": "Routes"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "rules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rules"
          },
          "blocked_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocked Domains"
          },
          "plan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "kind",
          "today",
          "trip",
          "routes",
          "topic",
          "instructions",
          "rules",
          "blocked_domains"
        ],
        "title": "RunContext"
      },
...
      "SuggestionBatchIn": {
        "properties": {
          "suggestions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Suggestions"
          }
        },
        "type": "object",
        "required": [
          "suggestions"
        ],
        "title": "SuggestionBatchIn"
      },
      "SuggestionBatchResult": {
        "properties": {
          "accepted": {
            "items": {
              "$ref": "#/components/schemas/AcceptedSuggestion"
            },
            "type": "array",
            "title": "Accepted"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/RejectedItem"
            },
            "type": "array",
            "title": "Rejected"
          },
          "duplicates": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Duplicates"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "rejected",
          "duplicates"
        ],
        "title": "SuggestionBatchResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}